Refactored stealth_helper.h, moved definitions to stealth_helper.c
